LPS-55928 move hits processors to hits package
