+++ portal-service/src/com/liferay/portal/kernel/search/hits/HitsProcessor.java
@@ -12,7 +12,11 @@
  * details.
  */
 
-package com.liferay.portal.kernel.search;
+package com.liferay.portal.kernel.search.hits;
+
+import com.liferay.portal.kernel.search.Hits;
+import com.liferay.portal.kernel.search.SearchContext;
+import com.liferay.portal.kernel.search.SearchException;
 
 /**
  * @author Michael C. Han
