+++ portal-service/src/com/liferay/portal/kernel/search/hits/packageinfo
@@ -1,0 +1,1 @@
+version 7.0.0
